requirements.txt
mcp 
python-dotenv 
nest_asyncio 
langchain_google_genai 
requests